Renamed Postbode.Client.PostbodeClient to Postbode.PostbodeClient. Added more configuration options
