+++ src/Postbode.Sendgrid/project.json
@@ -2,6 +2,7 @@
   "version": "1.0.0-*",
 
   "dependencies": {
+    "Microsoft.Extensions.Options": "1.0.0",
     "Postbode": "0.1.0-*",
     "System.Runtime.Serialization.Json": "4.0.2"
   },
