+++ src/Postbode/project.json
@@ -1,7 +1,9 @@
-﻿{
+{
   "version": "0.1.0-*",
 
-  "dependencies": {},
+  "dependencies": {
+    "Microsoft.Extensions.Options": "1.0.0"
+  },
 
   "frameworks": {
     "netstandard1.6": {
