+++ test/Postbode.Test/TestDeliveryService.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
-using Postbode.Client;
+using Postbode;
 using Postbode.Interfaces;
 
 namespace Postbode.Test
